Add JS Modules and button event listeners
Seperated large functions into different files and then imported
them into index.js.

Deleted most of teh content inside of template.html and because it is
all being created using JS.

Possible JS injections because of the use of inner.html however I don't
think that it is a problem.